+++ frontend/src/utils/utils.ts
@@ -394,6 +394,9 @@ export const prettyMilliseconds = function (n: number, options?: prettyMilliseco
             return "NaN";
         }
     }
+    else {
+        if (!isFinite(n)) return "N/A";
+    }
 
     // n is a finite number
     return prettyMillisecondsOriginal(n, options);
@@ -538,4 +541,22 @@ export class Message {
         this.duration = 3;
         this.update();
     }
+}
+
+
+//
+// https://stackoverflow.com/a/52171480/372434
+//
+const hashSeed = 0;
+export function hashString(str: string): number {
+    let h1 = 0xdeadbeef ^ hashSeed, h2 = 0x41c6ce57 ^ hashSeed;
+    for (let i = 0, ch; i < str.length; i++) {
+        ch = str.charCodeAt(i);
+        h1 = Math.imul(h1 ^ ch, 2654435761);
+        h2 = Math.imul(h2 ^ ch, 1597334677);
+    }
+    h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507) ^ Math.imul(h2 ^ (h2 >>> 13), 3266489909);
+    h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507) ^ Math.imul(h1 ^ (h1 >>> 13), 3266489909);
+    return 4294967296 * (2097151 & h2) + (h1 >>> 0);
+
 }